+++ anytype-cli/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog, and this project adheres to Semantic Versioning.
 
+## [Unreleased]
+
+- removed undocumented --keyring-service arg
+
 ## [0.2.3] 2026-01-12
 
 ### Changed